pyproject.toml
[tool.poetry]
name = "hikari-bot-template"
version = "0.1.0"
description = "simple template for hikari"
authors = ["OutofRange-Nico <outofrange210@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
hikari = "^2.0.0.dev120"
uvloop = "^0.17.0"
hikari-tanjun = "^2.16.0"
hikari-yuyo = "^1.14.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.280"
isort = "^5.12.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
